Fixed reactive navigation and implemented static preferences
Change bonusAmount type from Int to Long across features and services

+++ core/common/src/main/java/uz/yalla/client/core/common/button/BonusOverlay.kt
@@ -18,7 +18,7 @@ import uz.yalla.client.core.presentation.design.theme.YallaTheme
 
 @Composable
 fun BonusOverlay(
-    amount: Int,
+    amount: Long,
     onClick: () -> Unit
 ) {
     Card(
