+++ feature/bonus/presentation/src/main/java/uz/yalla/client/feature/bonus/bonus_account/view/BonusAccountScreen.kt
@@ -32,7 +32,7 @@ import uz.yalla.client.feature.bonus.presentation.R
 
 @Composable
 internal fun BonusAccountScreen(
-    balance: Int,
+    balance: Long,
     onIntent: (BonusAccountIntent) -> Unit
 ) {
     Scaffold(
